Move guest metrics gathering into subdir
Bug: 550013241
