+++ base/cvd/cuttlefish/host/libs/metrics/conversion/BUILD.bazel
@@ -30,8 +30,8 @@ cf_cc_library(
     deps = [
         "//cuttlefish/host/libs/config:data_image_policy",
         "//cuttlefish/host/libs/config:gpu_mode",
-        "//cuttlefish/host/libs/metrics:device_event_type",
-        "//cuttlefish/host/libs/metrics:guest_metrics",
+        "//cuttlefish/host/libs/metrics/guest:device_event_type",
+        "//cuttlefish/host/libs/metrics/guest:guest_metrics",
         "//external_proto:cf_flags_cc_proto",
         "//external_proto:cf_guest_cc_proto",
         "//external_proto:cf_metrics_event_v2_cc_proto",
@@ -61,11 +61,11 @@ cf_cc_library(
     srcs = ["metrics_data.cc"],
     hdrs = ["metrics_data.h"],
     deps = [
-        "//cuttlefish/host/libs/metrics/conversion:fetch_conversion",
         "//cuttlefish/host/libs/metrics:fetch_metrics",
+        "//cuttlefish/host/libs/metrics/conversion:fetch_conversion",
         "//cuttlefish/host/libs/metrics/conversion:guest_conversion",
-        "//cuttlefish/host/libs/metrics:guest_metrics",
         "//cuttlefish/host/libs/metrics/conversion:host_conversion",
+        "//cuttlefish/host/libs/metrics/guest:guest_metrics",
         "//cuttlefish/host/libs/metrics/host:host_metrics",
         "//external_proto:cf_log_cc_proto",
         "//external_proto:cf_metrics_event_v2_cc_proto",
