+++ base/cvd/cuttlefish/host/libs/metrics/conversion/guest_conversion.h
@@ -18,7 +18,7 @@
 
 #include <string_view>
 
-#include "cuttlefish/host/libs/metrics/guest_metrics.h"
+#include "cuttlefish/host/libs/metrics/guest/guest_metrics.h"
 #include "external_proto/cf_metrics_event_v2.pb.h"
 
 namespace cuttlefish {
