+++ base/cvd/cuttlefish/host/libs/metrics/conversion/metrics_data.h
@@ -22,7 +22,7 @@
 #include <vector>
 
 #include "cuttlefish/host/libs/metrics/fetch_metrics.h"
-#include "cuttlefish/host/libs/metrics/guest_metrics.h"
+#include "cuttlefish/host/libs/metrics/guest/guest_metrics.h"
 #include "cuttlefish/host/libs/metrics/host/host_metrics.h"
 #include "external_proto/cf_log.pb.h"
