+++ base/cvd/cuttlefish/host/libs/metrics/guest/device_event_type.h
