+++ base/cvd/cuttlefish/host/libs/metrics/guest/flag_metrics.cc
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-#include "cuttlefish/host/libs/metrics/flag_metrics.h"
+#include "cuttlefish/host/libs/metrics/guest/flag_metrics.h"
 
-#include "cuttlefish/host/libs/metrics/parsed_flags.h"
+#include "cuttlefish/host/libs/metrics/guest/parsed_flags.h"
 #include "cuttlefish/result/result.h"
 
 namespace cuttlefish {
