+++ base/cvd/cuttlefish/host/libs/metrics/guest/guest_metrics.cc
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "cuttlefish/host/libs/metrics/guest_metrics.h"
+#include "cuttlefish/host/libs/metrics/guest/guest_metrics.h"
 
 #include <string>
 #include <vector>
@@ -22,7 +22,7 @@
 #include "fmt/format.h"
 
 #include "cuttlefish/host/commands/assemble_cvd/boot_image_utils.h"
-#include "cuttlefish/host/libs/metrics/flag_metrics.h"
+#include "cuttlefish/host/libs/metrics/guest/flag_metrics.h"
 #include "cuttlefish/result/result.h"
 
 namespace cuttlefish {
